+++ hipd/modularization.h
@@ -10,18 +10,18 @@
 #ifndef HIP_HIPD_MODULARIZATION_H
 #define HIP_HIPD_MODULARIZATION_H
 
-int hip_register_handle_function(const uint32_t packet_type,
+int hip_register_handle_function(const uint8_t packet_type,
                                  const uint32_t ha_state,
-                                 int (*handle_function)(const uint32_t packet_type,
+                                 int (*handle_function)(const uint8_t packet_type,
                                                         const uint32_t ha_state,
                                                         struct hip_packet_context *ctx),
                                  const uint16_t priority);
 
-int hip_unregister_handle_function(const uint32_t packet_type,
+int hip_unregister_handle_function(const uint8_t packet_type,
                                    const uint32_t ha_state,
                                    const void *handle_function);
 
-int hip_run_handle_functions(const uint32_t packet_type,
+int hip_run_handle_functions(const uint8_t packet_type,
                              const uint32_t ha_state,
                              struct hip_packet_context *ctx);
